weighted corr fixad

+++ taxonomy_code/organisms_total.py
@@ -17,14 +17,13 @@ wastewater1 = "/storage/bergid/taxonomy_rewrites/taxonomy_all_ww_organisms_filte
 humangut = "/storage/bergid/taxonomy_rewrites/taxonomy_hg_organisms_filtered.tsv"
 outfile = "/storage/bergid/taxonomy_rewrites/taxonomy_all_organisms_filtered.tsv"
 
-
 df_hgut = pd.read_csv(humangut, sep="\t", index_col=0)
 df_waste1 = pd.read_csv(wastewater1, sep="\t", index_col=0)
 # df_waste2 = pd.read_csv(wastewater2, sep="\t", index_col=0)
 
 merged_df = pd.concat([df_waste1, df_hgut], axis=1, join="outer")
 merged_df = merged_df.fillna(0) # 1 for zinb, 0 for counts
-merged_df = merged_df.astype(int)
+# merged_df = merged_df.astype(int)
 merged_df.to_csv(outfile, sep="\t")
